move mutations up the dom
